Made GeoTiffReader and TiffTagReader available at the geotiff package level
