+++ raster/src/main/scala/geotrellis/raster/io/geotiff/MultibandGeoTiff.scala
@@ -18,7 +18,6 @@ package geotrellis.raster.io.geotiff
 
 import geotrellis.util.ByteReader
 import geotrellis.raster._
-import geotrellis.raster.io.geotiff.reader.GeoTiffReader
 import geotrellis.vector.Extent
 import geotrellis.proj4.CRS
 import geotrellis.raster.crop.Crop
